refactor: moved createDatasetSearchWhereExpression
