+++ sites/geohub/src/lib/helper/index.ts
@@ -1,6 +1,5 @@
 export * from './clean'
 export * from './clipSprite'
-export * from './createDatasetSearchWhereExpression'
 export * from './downloadFile'
 export * from './fetchUrl'
 export * from './fetchWithTimeout'
