+++ sites/geohub/src/routes/api/datasets/+server.ts
@@ -10,7 +10,7 @@ const connectionString = DATABASE_CONNECTION
 import { AccountSASPermissions, BlobServiceClient, StorageSharedKeyCredential } from '@azure/storage-blob'
 import { AZURE_STORAGE_ACCOUNT, AZURE_STORAGE_ACCESS_KEY } from '$lib/server/variables/private'
 import { TOKEN_EXPIRY_PERIOD_MSEC } from '$lib/constants'
-import { createDatasetSearchWhereExpression } from '$lib/helper/createDatasetSearchWhereExpression'
+import { createDatasetSearchWhereExpression } from '$lib/server/helpers/createDatasetSearchWhereExpression'
 
 /**
  * Datasets search API
