+++ sites/geohub/src/routes/api/tags/+server.ts
@@ -5,7 +5,7 @@ const { Pool } = pkg
 
 import { DATABASE_CONNECTION } from '$lib/server/variables/private'
 import type { Tag } from '$lib/types/Tag'
-import { createDatasetSearchWhereExpression } from '$lib/helper/createDatasetSearchWhereExpression'
+import { createDatasetSearchWhereExpression } from '$lib/server/helpers/createDatasetSearchWhereExpression'
 const connectionString = DATABASE_CONNECTION
 
 /**
